Fix broken Vte.Terminal.fork_command_full (Fedora 21)
Since Vte 0.38, Vte.Terminal.fork_command_full has been
deprecated. This is noticiable in sugar-build running in
Fedora 21.

The method is still present in previous supported versions
of Fedora.

Therefore, detect if the method is still present and use it, or,
in case is not present anymore, use spawn_async, which is its
equivalent in rewer versions.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>