examples: fix full screen issue
Specifically in the tree example, pygame canvas displays just black, and this happens because the screen is
not being updated every frame, unlike other examples.

This is probably related to a pygame issue https://github.com/pygame/pygame/issues/2538 as it happens if you run the example outside Pippy.

Issue was detected on F38 Gnome.